requirements.txt
# Controle de Vendas - Dependências Python
# Instale com: pip install -r requirements.txt

# ===========================
# Kivy - Framework UI
# ===========================
kivy==2.3.1

# ===========================
# Requisições HTTP
# ===========================
requests>=2.28.0

# ===========================
# Geração de PDF
# ===========================
reportlab>=4.0.0

# ===========================
# Geração de Imagens (para cupons)
# ===========================
Pillow>=10.0.0

# ===========================
# Segurança - Hash de senhas
# ===========================
bcrypt>=4.0.0

# ===========================
# Para Build do APK (Android)
# ===========================
# Descomente as linhas abaixo se estiver gerando APK
# pyjnius==1.4.5
# buildozer==1.5.0
# cython==0.29.35
